Use `#[cfg]` attributes to segment of Linux-specific code
Other platforms should now work, except when trying to bake `process`
recipes (but the code is now segmented to make adding support for
other platforms easier)
