+++ crates/brioche-packed-userland-exec/Cargo.toml
@@ -7,6 +7,9 @@ edition = "2021"
 bincode = "2.0.0-rc.3"
 brioche-pack = { path = "../brioche-pack", default-features = false }
 bstr = "1.8.0"
+cfg-if = "1.0.0"
 libc = "0.2.151"
 thiserror = "1.0.51"
+
+[target.'cfg(target_os = "linux")'.dependencies]
 userland-execve = "0.2.0"
