+++ crates/brioche/Cargo.toml
@@ -30,7 +30,6 @@ hex = "0.4.3"
 human-repr = "1.1.0"
 joinery = "3.1.0"
 json-canon = "0.1.3"
-libmount = "0.1.15"
 nix = { version = "0.27.1", features = ["user"] }
 opentelemetry = "0.21.0"
 opentelemetry-http = { version = "0.10.0", features = ["reqwest"] }
@@ -63,7 +62,6 @@ tracing = "0.1.40"
 tracing-opentelemetry = "0.22.0"
 tracing-subscriber = { version = "0.3.18", features = ["env-filter", "json", "tracing-log"] }
 ulid = "1.1.0"
-unshare = { git = "https://github.com/brioche-dev/unshare.git" }
 url = { version = "2.5.0", features = ["serde"] }
 urlencoding = "2.1.3"
 
@@ -74,6 +72,10 @@ mockito = "1.2.0"
 pretty_assertions = "1.4.0"
 tempdir = "0.3.7"
 zstd = "0.13.1"
+
+[target.'cfg(target_os = "linux")'.dependencies]
+libmount = "0.1.15"
+unshare = { git = "https://github.com/brioche-dev/unshare.git" }
 
 [[bench]]
 name = "resolve"
