+++ crates/brioche/tests/bake_process.rs
@@ -1,3 +1,5 @@
+#![cfg(target_os = "linux")]
+
 use std::collections::BTreeMap;
 
 use anyhow::Context;
